refactor: double 자료형 사용시 숫자 깨짐으로 인한 BigDecimal로 변경

+++ src/main/java/lotto/ouput/StatisticsOutput.java
@@ -1,5 +1,6 @@
 package lotto.ouput;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class StatisticsOutput {
@@ -13,7 +14,7 @@ public class StatisticsOutput {
         output.append("6개 일치 (2,000,000,000원) - "+winingHistory.get(3)+"개\n");
         System.out.print(output);
     }
-    public void outputProfit(double profit){
+    public void outputProfit(BigDecimal profit){
         System.out.println("총 수익률은 "+profit+"%입니다.");
     }
 }
